refactor(core): fold five copies of the store lock into one file_lock
kv.rs, overrides.rs, secrets.rs and barks.rs each carried a private
advisory lock type, and config_lock.rs carried a fifth as the public
ConfigLock. All five were the same thing: a blocking exclusive lock on a
sibling <name>.lock, through nix's Flock on unix and a share_mode(0) open
polled on a 2ms sleep against ERROR_SHARING_VIOLATION on Windows. The kv
and secrets blocks were 106 lines each and differed in two lines of doc
wording. That Windows retry loop existing five times is the sharp edge: a
fix to it had to land five times, on a platform the macOS gate never
compiles.

There is now one pub struct FileLock in shep-core, keyed on the path it
guards, and the four private types are gone.

Why a new module rather than atomic_file.rs: atomic_file is the replace
itself, and says so, "its first step and its last". The lock is not a step
of a replace. It is what keeps two replaces apart, and every caller holds
it across a read the replace never sees. Putting it there would have meant
widening that module's own doc to stop being true.

config_lock.rs keeps create_config_file, which is genuinely config
specific (it fixes the "shep" and ".toml.tmp" name pair that shep.toml and
dogs.toml share), and re-exports FileLock under the old ConfigLock name.
shep-core is published, and shep-daemon and shep-cli both name that type
across a crate boundary, so the alias keeps this non-breaking: no removed
public path, no changed signature, no version bump owed.

The module doc in each store explained the duplication as deliberate,
"copied rather than shared since KvLock is private to its module". That
reason held at two copies and stopped holding at three. Each now points at
crate::file_lock instead. config_lock.rs's own header said unifying the two
lock types was "a separate, later change"; this is that change.

No behaviour change. The concurrency tests that cover this, including
kv::tests::two_concurrent_writers_lose_nothing and its siblings in
overrides and barks, are untouched and green.

Gate, each run on its own:
- cargo fmt --all --check: 0
- cargo clippy --workspace --all-targets --all-features -- -D warnings: 0
- cargo test --workspace --all-features: 0, no failures across 40 suites
- RUSTDOCFLAGS="-D warnings" cargo doc --workspace --no-deps --all-features: 0
- cargo check --workspace --all-targets --all-features --target
  x86_64-pc-windows-gnu: 0, on its own CARGO_TARGET_DIR. This one matters
  here: the deduplicated Windows arm compiles nowhere else.
