+++ crates/shep-core/src/lib.rs
@@ -25,14 +25,16 @@
 pub mod atomic_file;
 pub mod barks;
 pub mod config;
-// The lock a config file's writers hold across their read-modify-write, and
-// the staging file they write through. Lives here rather than in shep-cli
-// (where it was born) so shep-daemon can hold it too, once it starts writing
-// `dogs.toml`.
+// The staging file `shep.toml` and `dogs.toml` are written through, and the
+// old name for the lock their writers hold. Lives here rather than in
+// shep-cli (where it was born) so shep-daemon can hold it too.
 pub mod config_lock;
 // The probe contract both sides of a dog's `--version`/`--schema` answer
 // parse: flag names, the answer grammar, the schema's secret marker key.
 pub mod dogs;
+// One advisory lock, keyed on the file it guards. Every store that
+// publishes a new value by `rename` holds it across the whole cycle.
+pub mod file_lock;
 pub mod kv;
 // One definition of the log-line timestamp for the writer and every reader:
 // the daemon stamps, and three different file readers in shep-cli strip.
